# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Documentation of To-do List App API.",
        "title": "To-do List App API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/login": {
            "post": {
                "description": "login user to get jwt token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Login User.",
                "parameters": [
                    {
                        "description": "JSON username and user_pass",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "create new user with username, user_email. user_pass.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create/Register New User.",
                "parameters": [
                    {
                        "description": "JSON username, user_email, and user_pass",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.CreateUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error insert user",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "docs.CreateUser": {
            "type": "object",
            "properties": {
                "user_email": {
                    "type": "string",
                    "example": "arifw.subagio17@gmail.com"
                },
                "user_pass": {
                    "type": "string",
                    "example": "rahasia87"
                },
                "username": {
                    "type": "string",
                    "example": "arifwidiasan"
                }
            }
        },
...
        "model.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "rahasia87"
                },
                "username": {
                    "type": "string",
                    "example": "arifwidiasan"
                }
            }
        }
    }
}